Applying a patch to compute index for system locks

+++ src/core/system.hpp
@@ -441,7 +441,7 @@ inline void System::registerPluginOption ( const std::string &option, const std:
 
 inline int System::nextThreadId () { return _threadIdSeed++; }
 
-inline Lock * System::getLockAddress ( void *addr ) const { return &_lockPool[((((long long)addr)>>8)%_lockPoolSize)];} ;
+inline Lock * System::getLockAddress ( void *addr ) const { return &_lockPool[((((uintptr_t)addr)>>8)%_lockPoolSize)];} ;
 
 inline bool System::dlbEnabled() const { return _enable_dlb; }
